+++ test/test-react-conveyor.jsx
@@ -17,11 +17,11 @@ describe('ReactConveyor', function() {
 
   it('exposes correct props for initial render', function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     // Use `shallow` as this will not mount and not trigger the
     // `componentDidMount` lifecycle hook.
-    const wrapper = shallow(<Conveyor fetch={fetch}>{children}</Conveyor>);
+    const wrapper = shallow(<Conveyor fields={fields}>{children}</Conveyor>);
 
     expect(wrapper.props()).toEqual({
       bar: undefined,
@@ -33,28 +33,28 @@ describe('ReactConveyor', function() {
     });
   });
 
-  it('calls all fragments on mounting', function() {
+  it('calls all fields on mounting', function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
     const extraProps = {
       someProp: 1,
       someOtherProp: true,
     };
-    mount(<Conveyor {...extraProps} fetch={fetch}>{children}</Conveyor>);
+    mount(<Conveyor {...extraProps} fields={fields}>{children}</Conveyor>);
 
-    expect(fetch.foo.calledOnce).toBe(true);
-    expect(fetch.foo.args[0]).toEqual([extraProps]);
+    expect(fields.foo.calledOnce).toBe(true);
+    expect(fields.foo.args[0]).toEqual([extraProps]);
 
-    expect(fetch.bar.calledOnce).toBe(true);
-    expect(fetch.bar.args[0]).toEqual([extraProps]);
+    expect(fields.bar.calledOnce).toBe(true);
+    expect(fields.bar.args[0]).toEqual([extraProps]);
   });
 
   it('forwards extra props to children', function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     const wrapper = mount(
-      <Conveyor someProp={1} someOtherProp fetch={fetch}>{children}</Conveyor>
+      <Conveyor someProp={1} someOtherProp fields={fields}>{children}</Conveyor>
     );
 
     const props = wrapper.find(CustomChild).first().props();
@@ -64,9 +64,9 @@ describe('ReactConveyor', function() {
 
   it('exposes correct props after mounting', function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
-    const wrapper = mount(<Conveyor fetch={fetch}>{children}</Conveyor>);
+    const wrapper = mount(<Conveyor fields={fields}>{children}</Conveyor>);
 
     expect(wrapper.find(CustomChild).first().props()).toEqual({
       bar: undefined,
@@ -78,15 +78,15 @@ describe('ReactConveyor', function() {
     });
   });
 
-  it('exposes correct props after fragments have resolved / rejected', async function() {
+  it('exposes correct props after fields have resolved / rejected', async function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
-    const wrapper = mount(<Conveyor fetch={fetch}>{children}</Conveyor>);
+    const fields = promiseFactories();
+    const wrapper = mount(<Conveyor fields={fields}>{children}</Conveyor>);
 
     const fooValue = {};
     const barError = new Error('Bar');
-    fetch.foo.resolve(0, fooValue);
-    fetch.bar.reject(0, barError);
+    fields.foo.resolve(0, fooValue);
+    fields.bar.reject(0, barError);
 
     await helpers.sleep(0);
     wrapper.update(); // External async update so we need to force re-render.
@@ -101,15 +101,15 @@ describe('ReactConveyor', function() {
     expect(props.reload).toBe(wrapper.instance().reload);
   });
 
-  it('reloads fragments when props change', function() {
+  it('reloads fields when props change', function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = {foo: helpers.controlledPromiseFactory()};
-    const wrapper = mount(<Conveyor someProp={1} fetch={fetch}>{children}</Conveyor>);
+    const fields = {foo: helpers.controlledPromiseFactory()};
+    const wrapper = mount(<Conveyor someProp={1} fields={fields}>{children}</Conveyor>);
 
     wrapper.setProps({someProp: 2});
 
-    expect(fetch.foo.calledTwice).toBe(true);
-    expect(fetch.foo.args).toEqual([
+    expect(fields.foo.calledTwice).toBe(true);
+    expect(fields.foo.args).toEqual([
       [ { someProp: 1 } ],
       [ { someProp: 2 } ]
     ]);
@@ -117,12 +117,12 @@ describe('ReactConveyor', function() {
 
   it('resolves race conditions when promises resolve in the wrong order', async function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = {foo: helpers.controlledPromiseFactory()};
-    const wrapper = mount(<Conveyor someProp={1} fetch={fetch}>{children}</Conveyor>);
+    const fields = {foo: helpers.controlledPromiseFactory()};
+    const wrapper = mount(<Conveyor someProp={1} fields={fields}>{children}</Conveyor>);
 
     wrapper.setProps({someProp: 2});
 
-    fetch.foo.resolve(1, 1);
+    fields.foo.resolve(1, 1);
 
     await helpers.sleep(0);
     wrapper.update(); // External async update so we need to force re-render.
@@ -132,7 +132,7 @@ describe('ReactConveyor', function() {
     expect(props.foo).toBe(1);
     expect(props.errors).toBe(null);
 
-    fetch.foo.reject(0, new Error('foo'));
+    fields.foo.reject(0, new Error('foo'));
 
     await helpers.sleep(0);
     wrapper.update(); // External async update so we need to force re-render.
@@ -141,9 +141,9 @@ describe('ReactConveyor', function() {
     expect(props.errors).toBe(null);
   });
 
-  it('honors `mapPropsToArgs` when calling fragments', function() {
+  it('honors `mapPropsToArgs` when calling fields', function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     const mapPropsToArgs = {
       foo: x => ({value: x.fooInputValue}),
@@ -156,21 +156,21 @@ describe('ReactConveyor', function() {
     };
 
     mount(
-      <Conveyor {...extraProps} fetch={fetch} mapPropsToArgs={mapPropsToArgs}>
+      <Conveyor {...extraProps} fields={fields} mapPropsToArgs={mapPropsToArgs}>
         {children}
       </Conveyor>
     );
 
-    expect(fetch.foo.calledOnce).toBe(true);
-    expect(fetch.foo.args[0]).toEqual([{value: 1}]);
+    expect(fields.foo.calledOnce).toBe(true);
+    expect(fields.foo.args[0]).toEqual([{value: 1}]);
 
-    expect(fetch.bar.calledOnce).toBe(true);
-    expect(fetch.bar.args[0]).toEqual([{value: 2}]);
+    expect(fields.bar.calledOnce).toBe(true);
+    expect(fields.bar.args[0]).toEqual([{value: 2}]);
   });
 
   it('honors `mapPropsToArgs` when receiving props', async function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     const mapPropsToArgs = {
       foo: x => ({value: x.fooInputValue}),
@@ -183,28 +183,28 @@ describe('ReactConveyor', function() {
     };
 
     const wrapper = mount(
-      <Conveyor {...extraProps} fetch={fetch} mapPropsToArgs={mapPropsToArgs}>
+      <Conveyor {...extraProps} fields={fields} mapPropsToArgs={mapPropsToArgs}>
         {children}
       </Conveyor>
     );
 
     wrapper.setProps({barInputValue: 3});
 
-    expect(fetch.foo.calledOnce).toBe(true); // Did not reload
-    expect(fetch.foo.args[0]).toEqual([{value: 1}]);
+    expect(fields.foo.calledOnce).toBe(true); // Did not reload
+    expect(fields.foo.args[0]).toEqual([{value: 1}]);
 
-    expect(fetch.bar.calledTwice).toBe(true);
-    expect(fetch.bar.args).toEqual([
+    expect(fields.bar.calledTwice).toBe(true);
+    expect(fields.bar.args).toEqual([
       [{value: 2}],
       [{value: 3}],
     ]);
   });
 
-  it('only reloads affected fragments when `mapPropsToArgs` return value changes', function() {
+  it('only reloads affected fields when `mapPropsToArgs` return value changes', function() {
     // Had a case where reloading when the function changed but the result
     // didn't was undesirable.
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     const mapPropsToArgs = {
       foo: x => ({value: x.fooInputValue}),
@@ -217,7 +217,7 @@ describe('ReactConveyor', function() {
     };
 
     const wrapper = mount(
-      <Conveyor {...extraProps} fetch={fetch} mapPropsToArgs={mapPropsToArgs}>
+      <Conveyor {...extraProps} fields={fields} mapPropsToArgs={mapPropsToArgs}>
         {children}
       </Conveyor>
     );
@@ -229,11 +229,11 @@ describe('ReactConveyor', function() {
 
     wrapper.setProps({mapPropsToArgs: nextMapPropsToArgs});
 
-    expect(fetch.foo.calledOnce).toBe(true); // Did not reload
-    expect(fetch.foo.args[0]).toEqual([{value: 1}]);
+    expect(fields.foo.calledOnce).toBe(true); // Did not reload
+    expect(fields.foo.args[0]).toEqual([{value: 1}]);
 
-    expect(fetch.bar.calledTwice).toBe(true);
-    expect(fetch.bar.args).toEqual([
+    expect(fields.bar.calledTwice).toBe(true);
+    expect(fields.bar.args).toEqual([
       [{value: 2}],
       [{value: 3}],
     ]);
@@ -242,101 +242,102 @@ describe('ReactConveyor', function() {
 
   it('auto-refreshes globally when provided a number', async function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     mount(
-      <Conveyor fetch={fetch} refresh={20}>
+      <Conveyor fields={fields} refresh={20}>
         {children}
       </Conveyor>
     );
 
-    fetch.foo.resolve(0, 1);
-    fetch.bar.resolve(0, 2);
+    fields.foo.resolve(0, 1);
+    fields.bar.resolve(0, 2);
 
     await helpers.sleep(30);
 
-    expect(fetch.foo.calledTwice).toBe(true);
-    expect(fetch.bar.calledTwice).toBe(true);
+    expect(fields.foo.calledTwice).toBe(true);
+    expect(fields.bar.calledTwice).toBe(true);
 
-    fetch.foo.resolve(1, 1);
-    fetch.bar.resolve(1, 2);
+    fields.foo.resolve(1, 1);
+    fields.bar.resolve(1, 2);
 
     await helpers.sleep(30);
 
-    expect(fetch.foo.calledThrice).toBe(true);
-    expect(fetch.bar.calledThrice).toBe(true);
+    expect(fields.foo.calledThrice).toBe(true);
+    expect(fields.bar.calledThrice).toBe(true);
   });
 
-  it('auto-refreshes per fragment when provided an object', async function() {
+  it('auto-refreshes per field when provided an object', async function() {
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     mount(
-      <Conveyor fetch={fetch} refresh={{foo: 20}}>
+      <Conveyor fields={fields} refresh={{foo: 20}}>
         {children}
       </Conveyor>
     );
 
-    fetch.foo.resolve(0, 1);
-    fetch.bar.resolve(0, 2);
+    fields.foo.resolve(0, 1);
+    fields.bar.resolve(0, 2);
 
     await helpers.sleep(30);
 
-    expect(fetch.foo.calledTwice).toBe(true);
-    expect(fetch.bar.calledOnce).toBe(true);
+    expect(fields.foo.calledTwice).toBe(true);
+    expect(fields.bar.calledOnce).toBe(true);
 
-    fetch.foo.resolve(1, 1);
+    fields.foo.resolve(1, 1);
 
     await helpers.sleep(30);
 
-    expect(fetch.foo.calledThrice).toBe(true);
-    expect(fetch.bar.calledOnce).toBe(true);
+    expect(fields.foo.calledThrice).toBe(true);
+    expect(fields.bar.calledOnce).toBe(true);
   });
 
-  it('does not auto-refresh until fragment resolves', async function() {
+  it('does not auto-refresh until field resolves', async function() {
     // Avoid unnecessary race-conditions and function calls
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     mount(
-      <Conveyor fetch={fetch} refresh={{foo: 20}}>
+      <Conveyor fields={fields} refresh={{foo: 20}}>
         {children}
       </Conveyor>
     );
 
     await helpers.sleep(30);
 
-    expect(fetch.foo.calledOnce).toBe(true);
-    fetch.foo.resolve(0, 1);
+    expect(fields.foo.calledOnce).toBe(true);
+    fields.foo.resolve(0, 1);
 
     await helpers.sleep(30);
 
-    expect(fetch.foo.calledTwice).toBe(true);
+    expect(fields.foo.calledTwice).toBe(true);
   });
 
-  it('does not auto-refresh on fragment rejection', async function() {
+  it('does not auto-refresh on field rejection', async function() {
     // Avoid unnecessary race-conditions and function calls
     const children = props => <CustomChild {...props}/>;
-    const fetch = promiseFactories();
+    const fields = promiseFactories();
 
     mount(
-      <Conveyor fetch={fetch} refresh={{foo: 20}}>
+      <Conveyor fields={fields} refresh={{foo: 20}}>
         {children}
       </Conveyor>
     );
 
-    fetch.foo.reject(0, new Error('Foo failed.'));
+    fields.foo.reject(0, new Error('Foo failed.'));
 
     await helpers.sleep(30);
 
-    expect(fetch.foo.calledOnce).toBe(true);
+    expect(fields.foo.calledOnce).toBe(true);
   });
 });
 
 describe('ReactConveyor.wrapComponent', function() {
+
   it('returns a Component function that renders ReactConveyor correctly', function() {
     const defaultProps = {
-      fetch: promiseFactories(),
+      fields: promiseFactories(),
       mapPropsToArgs: {
         foo: props => props.fooInput,
       },
@@ -362,4 +363,5 @@ describe('ReactConveyor.wrapComponent', function() {
       fooInput: 1,
     });
   });
+
 });