+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes with version x.x.x
 - Change `prepublish` script to `prepublishOnly`.  
   See: https://docs.npmjs.com/misc/scripts#prepublish-and-prepare.
 - Do not prevent fragments from updating when one promise is already in fight.
+- **Breaking** Rename `fetch` prop to `fields` + all internal mention of `fragment` to `field`.
 
 
 Changes with version 0.0.1